move metric service to commons
