+++ commons/src/main/java/org/eclipse/kapua/commons/metric/MetricsService.java
@@ -9,7 +9,7 @@
  * Contributors:
  *     Eurotech - initial API and implementation
  *******************************************************************************/
-package org.eclipse.kapua.service.metric;
+package org.eclipse.kapua.commons.metric;
 
 import org.eclipse.kapua.KapuaException;
 import org.eclipse.kapua.service.KapuaService;
