+++ commons/src/main/java/org/eclipse/kapua/commons/metric/MetricsServiceImpl.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2011, 2016 Eurotech and/or its affiliates and others
+ * Copyright (c) 2011, 2017 Eurotech and/or its affiliates and others
  *
  * All rights reserved. This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License v1.0
@@ -17,8 +17,6 @@ import java.util.concurrent.TimeUnit;
 import org.eclipse.kapua.KapuaException;
 import org.eclipse.kapua.commons.setting.system.SystemSetting;
 import org.eclipse.kapua.commons.setting.system.SystemSettingKey;
-import org.eclipse.kapua.locator.KapuaProvider;
-import org.eclipse.kapua.service.metric.MetricsService;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -36,7 +34,6 @@ import com.codahale.metrics.Timer;
  *
  * @since 1.0
  */
-@KapuaProvider
 public class MetricsServiceImpl implements MetricsService {
 
     private static Logger logger = LoggerFactory.getLogger(MetricsServiceImpl.class);
@@ -50,7 +47,7 @@ public class MetricsServiceImpl implements MetricsService {
     /**
      * Default metric service constructor
      */
-    public MetricsServiceImpl() {
+    MetricsServiceImpl() {
         metricRegistry = new MetricRegistry();
 
         if (isJmxEnabled()) {
